Rename `RemoveSetterGetterDocblockFixer` to `RemoveMethodNameDuplicateDescriptionFixer`
