+++ tests/Fixer/Annotation/RemoveMethodNameDuplicateDescriptionFixer/Fixture/with_return_date_time.php.inc
@@ -1,6 +1,6 @@
 <?php
 
-namespace Symplify\CodingStandard\Tests\Fixer\Annotation\RemoveSetterGetterDocblockFixer\Fixture;
+namespace Symplify\CodingStandard\Tests\Fixer\Annotation\RemoveMethodNameDuplicateDescriptionFixer\Fixture;
 
 final class WithReturnDateTime
 {
@@ -18,7 +18,7 @@ final class WithReturnDateTime
 -----
 <?php
 
-namespace Symplify\CodingStandard\Tests\Fixer\Annotation\RemoveSetterGetterDocblockFixer\Fixture;
+namespace Symplify\CodingStandard\Tests\Fixer\Annotation\RemoveMethodNameDuplicateDescriptionFixer\Fixture;
 
 final class WithReturnDateTime
 {
