+++ tests/Fixer/Annotation/RemoveMethodNameDuplicateDescriptionFixer/config/configured_rule.php
@@ -2,11 +2,11 @@
 
 declare(strict_types=1);
 
-use Symplify\CodingStandard\Fixer\Annotation\RemoveSetterGetterDocblockFixer;
+use Symplify\CodingStandard\Fixer\Annotation\RemoveMethodNameDuplicateDescriptionFixer;
 use Symplify\EasyCodingStandard\Config\ECSConfig;
 
 return static function (ECSConfig $ecsConfig): void {
     $ecsConfig->rules([
-        RemoveSetterGetterDocblockFixer::class,
+        RemoveMethodNameDuplicateDescriptionFixer::class,
     ]);
 };
